+++ tasks/discover-pdrs/index.js
@@ -1,91 +1,61 @@
 'use strict';
 
-const cumulusMessageAdapter = require('@cumulus/cumulus-message-adapter-js');
 const get = require('lodash.get');
-const pdr = require('@cumulus/ingest/pdr');
-const errors = require('@cumulus/errors');
-const log = require('@cumulus/common/log');
+const pFilter = require('p-filter');
+const { s3ObjectExists } = require('@cumulus/aws-client/S3');
+const { buildProviderClient } = require('@cumulus/ingest/providerClientUtils');
+const { normalizeProviderPath } = require('@cumulus/ingest/util');
+const { runCumulusTask } = require('@cumulus/cumulus-message-adapter-js');
+
+/**
+ * Fetch a list of files from the provider
+ *
+ * @param {Object} providerConfig - the connection config for the provider
+ * @param {bool} useList - flag to tell ftp server to use 'LIST' instead of 'STAT'
+ * @param {*} path - the provider path to search
+ * @returns {Array<Object>} a list of discovered file objects
+ */
+const listFiles = (providerConfig, useList, path) =>
+  buildProviderClient({ ...providerConfig, useList }).list(path);
+
+const isPdrFile = ({ name }) => name.toUpperCase().endsWith('.PDR');
+
+const isNewPdr = (bucket, stackName, folder, pdr) =>
+  s3ObjectExists({
+    Bucket: bucket,
+    Key: `${stackName}/${folder}/${pdr.name}`
+  }).then((pdrIsInS3) => pdrIsInS3 === false);
 
 /**
  * Discover PDRs
  *
  * @param {Object} event - a simplified Cumulus event with input and config properties
- * @returns {Promise.<Array>} - resolves to an array describing PDRs
+ * @returns {Promise<Array>} - resolves to an array describing PDRs
  */
-function discoverPdrs(event) {
-  try {
-    const config = get(event, 'config', {});
-    const stack = config.stack;
-    const bucket = config.bucket;
-    const collection = config.collection;
-    const provider = config.provider;
-    const providerPath = config.provider_path || collection.provider_path;
-    const filterPdrs = config.filterPdrs || null;
+const discoverPdrs = async ({ config }) => {
+  const discoveredFiles = await listFiles(
+    config.provider,
+    config.useList,
+    normalizeProviderPath(config.collection.provider_path)
+  );
 
-    // FIXME Can config.folder not be used?
+  const discoveredPdrs = discoveredFiles.filter(isPdrFile);
 
-    log.info('Received the provider', { provider: get(provider, 'id') });
-
-    const Discover = pdr.selector('discover', provider.protocol);
-    const discover = new Discover(
-      stack,
-      bucket,
-      providerPath,
-      provider,
-      config.useList,
-      'pdrs',
-      config.force || false
+  let pdrsToReturn;
+  if (get(config, 'force', false)) {
+    pdrsToReturn = discoveredPdrs;
+  } else {
+    pdrsToReturn = await pFilter(
+      discoveredPdrs,
+      (pdr) => isNewPdr(config.bucket, config.stack, 'pdrs', pdr)
     );
-
-    log.debug('Starting PDR discovery');
-
-    return discover.discover()
-      .then((pdrs) => {
-        if (discover.connected) discover.end();
-
-        // filter pdrs using filterPDrs
-        if (filterPdrs && pdrs.length > 0) {
-          log.info(`Filtering ${pdrs.length} with ${filterPdrs}`);
-          const fpdrs = pdrs.filter((p) => p.name.match(filterPdrs));
-          return { pdrs: fpdrs };
-        }
-
-        return { pdrs };
-      })
-      .catch((e) => {
-        log.error(e);
-
-        if (discover.connected) {
-          discover.end();
-          log.debug(`Ending ${provider.protocol} connection`);
-        }
-
-        if (e.toString().includes('ECONNREFUSED')) {
-          const err = new errors.RemoteResourceError('Connection Refused');
-          log.error(err);
-          throw err;
-        } else if (e.message.includes('Please login with USER and PASS')) {
-          const err = new errors.FTPError('Login incorrect');
-          log.error(err);
-          throw err;
-        } else if (e.details && e.details.status === 'timeout') {
-          const err = new errors.ConnectionTimeout('connection Timed out');
-          log.error(err);
-          throw err;
-        } else if (e.details && e.details.status === 'notfound') {
-          const err = new errors.HostNotFound(`${e.details.url} not found`);
-          log.error(err);
-          throw err;
-        }
-
-        throw e;
-      });
-  } catch (e) {
-    log.error(e);
-    throw e;
   }
-}
-exports.discoverPdrs = discoverPdrs; // exported to support testing
+
+  const pdrNameFilter = get(config, 'filterPdrs', /.*/);
+  return {
+    pdrs: pdrsToReturn.filter(({ name }) => name.match(pdrNameFilter))
+  };
+};
 
 /**
  * Lambda handler
@@ -95,7 +65,11 @@ exports.discoverPdrs = discoverPdrs; // exported to support testing
  * @param {Function} callback - an AWS Lambda handler
  * @returns {undefined} - does not return a value
  */
-function handler(event, context, callback) {
-  cumulusMessageAdapter.runCumulusTask(discoverPdrs, event, context, callback);
-}
-exports.handler = handler;
+const handler = (event, context, callback) => {
+  runCumulusTask(discoverPdrs, event, context, callback);
+};
+
+module.exports = {
+  discoverPdrs,
+  handler
+};